move PhysicsTick to own file
